# contracts/openapi/endpoint-platform-v1.yaml
openapi: "3.1.0"
info:
  title: "Endpoint Platform API"
  version: "v1"
paths:
...
  "/api/v1/devices/{device_id}/module-operations":
    post:
      tags:
        - "endpoint-module-execution"
      summary: "Create Module Operation"
      description: "Create/replay a module parent; only Gateway WSS can deliver its children."
      operationId: "create_module_operation_api_v1_devices__device_id__module_operations_post"
      parameters:
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "Idempotency-Key"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 8
            maxLength: 128
            pattern: "^[!-~][ -~]{6,126}[!-~]$"
            title: "Idempotency-Key"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/ModuleOperationCreateV1"
      responses:
        "201":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/ModuleOperationEnvelope"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/module-operations/{operation_id}":
    get:
      tags:
        - "endpoint-module-execution"
      summary: "Read Module Operation"
      description: "Read only a caller-owned parent and its safe typed child summaries."
      operationId: "read_module_operation_api_v1_module_operations__operation_id__get"
      parameters:
        -
          name: "operation_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Operation Id"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/ModuleOperationDetailEnvelope"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
  "/api/v1/modules/{module_key}":
    get:
      tags:
        - "endpoint-modules"
      summary: "Read Module"
      operationId: "read_module_api_v1_modules__module_key__get"
      parameters:
        -
          name: "module_key"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Module Key"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/ModuleVersionViewEnvelope"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
  "/api/v1/modules/{module_key}/versions/{version}/lab-operations/{device_id}":
    post:
      tags:
        - "endpoint-module-execution"
      summary: "Create Module Lab Operation"
      description: "Create a lab-only module parent while its immutable version is validated."
      operationId: "create_module_lab_operation_api_v1_modules__module_key__versions__version__lab_operations__device_id__post"
      parameters:
        -
          name: "module_key"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Module Key"
        -
          name: "version"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Version"
        -
          name: "device_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Device Id"
        -
          name: "Idempotency-Key"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 8
            maxLength: 128
            pattern: "^[!-~][ -~]{6,126}[!-~]$"
            title: "Idempotency-Key"
        -
          name: "X-Correlation-ID"
          in: "header"
          required: true
          schema:
            type: "string"
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            title: "X-Correlation-Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/ModuleLabOperationCreateV1"
      responses:
        "201":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/ModuleOperationEnvelope"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    ModuleLabOperationCreateV1:
      additionalProperties: false
      description: "Inputs for a validated-only Endpoint-owned module lab parent."
      properties:
        schema_version:
          const: "endpoint_module_lab_operation_create_v1"
          title: "Schema Version"
          type: "string"
        inputs:
          maxProperties: 8
          minProperties: 0
          patternProperties:
            "^[a-z][a-z0-9_]{0,63}$":
              anyOf:
                -
                  type: "string"
                -
                  type: "integer"
          propertyNames:
            maxLength: 64
            minLength: 1
          title: "Inputs"
          type: "object"
      required:
        - "schema_version"
        - "inputs"
      title: "ModuleLabOperationCreateV1"
      type: "object"
...
    ModuleOperationCreateV1:
      additionalProperties: false
      properties:
        schema_version:
          const: "endpoint_module_operation_create_v1"
          title: "Schema Version"
          type: "string"
        module_key:
          maxLength: 128
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*){1,7}$"
          title: "Module Key"
          type: "string"
        version:
          maxLength: 64
          minLength: 5
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          title: "Version"
          type: "string"
        inputs:
          maxProperties: 8
          minProperties: 0
          patternProperties:
            "^[a-z][a-z0-9_]{0,63}$":
              anyOf:
                -
                  type: "string"
                -
                  type: "integer"
          propertyNames:
            maxLength: 64
            minLength: 1
          title: "Inputs"
          type: "object"
      required:
        - "schema_version"
        - "module_key"
        - "version"
        - "inputs"
      title: "ModuleOperationCreateV1"
      type: "object"
    ModuleOperationV1:
      additionalProperties: false
      properties:
        schema_version:
          const: "endpoint_module_operation_v1"
          title: "Schema Version"
          type: "string"
        operation_id:
          format: "uuid"
          title: "Operation Id"
          type: "string"
        device_id:
          format: "uuid"
          title: "Device Id"
          type: "string"
        module_key:
          maxLength: 128
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*){1,7}$"
          title: "Module Key"
          type: "string"
        version:
          maxLength: 64
          minLength: 5
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          title: "Version"
          type: "string"
        status:
          enum:
            - "queued"
            - "delivered"
            - "acknowledged"
            - "running"
            - "succeeded"
            - "failed"
            - "canceled"
            - "expired"
          title: "Status"
          type: "string"
        created_at:
          format: "date-time"
          title: "Created At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        deadline_at:
          format: "date-time"
          title: "Deadline At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        completed_at:
          anyOf:
            -
              format: "date-time"
              type: "string"
              "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
            -
              type: "null"
          default: null
          title: "Completed At"
      required:
        - "schema_version"
        - "operation_id"
        - "device_id"
        - "module_key"
        - "version"
        - "status"
        - "created_at"
        - "deadline_at"
      title: "ModuleOperationV1"
      type: "object"
    ModuleOperationDetailV1:
      additionalProperties: false
      properties:
        schema_version:
          const: "endpoint_module_operation_v1"
          title: "Schema Version"
          type: "string"
        operation_id:
          format: "uuid"
          title: "Operation Id"
          type: "string"
        device_id:
          format: "uuid"
          title: "Device Id"
          type: "string"
        module_key:
          maxLength: 128
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*){1,7}$"
          title: "Module Key"
          type: "string"
        version:
          maxLength: 64
          minLength: 5
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          title: "Version"
          type: "string"
        status:
          enum:
            - "queued"
            - "delivered"
            - "acknowledged"
            - "running"
            - "succeeded"
            - "failed"
            - "canceled"
            - "expired"
          title: "Status"
          type: "string"
        created_at:
          format: "date-time"
          title: "Created At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        deadline_at:
          format: "date-time"
          title: "Deadline At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
        completed_at:
          anyOf:
            -
              format: "date-time"
              type: "string"
              "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
            -
              type: "null"
          default: null
          title: "Completed At"
        expected_step_count:
          maximum: 8
          minimum: 1
          title: "Expected Step Count"
          type: "integer"
        steps:
          items:
            "$ref": "#/components/schemas/ModuleOperationDetailV1__ModuleOperationStepV1"
          maxItems: 8
          minItems: 1
          title: "Steps"
          type: "array"
      required:
        - "schema_version"
        - "operation_id"
        - "device_id"
        - "module_key"
        - "version"
        - "status"
        - "created_at"
        - "deadline_at"
        - "expected_step_count"
        - "steps"
      title: "ModuleOperationDetailV1"
      type: "object"
...
    ModuleOperationDetailV1__DnsResolveResultV1:
      additionalProperties: false
      properties:
        schema_version:
          const: "dns_resolve_result_v1"
          title: "Schema Version"
          type: "string"
        target:
          maxLength: 253
          minLength: 1
          title: "Target"
          type: "string"
        canonical_name:
          anyOf:
            -
              maxLength: 253
              minLength: 1
              type: "string"
            -
              type: "null"
          default: null
          title: "Canonical Name"
        addresses:
          items:
            "$ref": "#/components/schemas/ModuleOperationDetailV1__NetworkAddressV1"
          maxItems: 16
          title: "Addresses"
          type: "array"
        address_count:
          maximum: 16
          minimum: 0
          title: "Address Count"
          type: "integer"
        status:
          enum:
            - "succeeded"
            - "failed"
          title: "Status"
          type: "string"
        error_code:
          anyOf:
            -
              maxLength: 64
              minLength: 1
              pattern: "^[a-z][a-z0-9_]{0,63}$"
              type: "string"
            -
              type: "null"
          default: null
          title: "Error Code"
        collected_at:
          format: "date-time"
          title: "Collected At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
      required:
        - "schema_version"
        - "target"
        - "address_count"
        - "status"
        - "collected_at"
      title: "DnsResolveResultV1"
      type: "object"
    ModuleOperationDetailV1__ModuleOperationStepV1:
      additionalProperties: false
      properties:
        sequence:
          maximum: 7
          minimum: 0
          title: "Sequence"
          type: "integer"
        capability:
          enum:
            - "dns.resolve"
            - "network.ping"
            - "tcp.connect"
            - "route.get"
            - "adapter.list"
            - "system.service_status"
          title: "Capability"
          type: "string"
        status:
          enum:
            - "queued"
            - "delivered"
            - "acknowledged"
            - "running"
            - "succeeded"
            - "failed"
            - "canceled"
            - "expired"
          title: "Status"
          type: "string"
        error_code:
          anyOf:
            -
              maxLength: 64
              minLength: 3
              pattern: "^[a-z][a-z0-9_]{1,63}$"
              type: "string"
            -
              type: "null"
          default: null
          title: "Error Code"
        safe_result:
          anyOf:
            -
              discriminator:
                mapping:
                  adapter_list_result_v1: "#/$defs/AdapterListResultV1"
                  dns_resolve_result_v1: "#/$defs/DnsResolveResultV1"
                  network_ping_result_v1: "#/$defs/NetworkPingResultV1"
                  route_get_result_v1: "#/$defs/RouteGetResultV1"
                  service_status_result_v1: "#/$defs/ServiceStatusResultV1"
                  tcp_connect_result_v1: "#/$defs/TcpConnectResultV1"
                propertyName: "schema_version"
              oneOf:
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__DnsResolveResultV1"
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__NetworkPingResultV1"
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__TcpConnectResultV1"
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__RouteGetResultV1"
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__AdapterListResultV1"
                -
                  "$ref": "#/components/schemas/ModuleOperationDetailV1__ServiceStatusResultV1"
            -
              type: "null"
          default: null
          title: "Safe Result"
      required:
        - "sequence"
        - "capability"
        - "status"
      title: "ModuleOperationStepV1"
      type: "object"
...
    ModuleOperationDetailV1__TcpConnectResultV1:
      additionalProperties: false
      properties:
        schema_version:
          const: "tcp_connect_result_v1"
          title: "Schema Version"
          type: "string"
        target:
          maxLength: 253
          minLength: 1
          title: "Target"
          type: "string"
        resolved_ip:
          anyOf:
            -
              maxLength: 253
              minLength: 1
              type: "string"
            -
              type: "null"
          default: null
          title: "Resolved Ip"
        port:
          maximum: 65535
          minimum: 1
          title: "Port"
          type: "integer"
        reachable:
          title: "Reachable"
          type: "boolean"
        latency_ms:
          anyOf:
            -
              maximum: 60000
              minimum: 0
              type: "number"
            -
              type: "null"
          default: null
          title: "Latency Ms"
        status:
          enum:
            - "succeeded"
            - "failed"
          title: "Status"
          type: "string"
        error_code:
          anyOf:
            -
              maxLength: 64
              minLength: 1
              pattern: "^[a-z][a-z0-9_]{0,63}$"
              type: "string"
            -
              type: "null"
          default: null
          title: "Error Code"
        collected_at:
          format: "date-time"
          title: "Collected At"
          type: "string"
          "$comment": "model-only UTC normalization: timezone-aware values are normalized to UTC."
      required:
        - "schema_version"
        - "target"
        - "port"
        - "reachable"
        - "status"
        - "collected_at"
      title: "TcpConnectResultV1"
      type: "object"
...
    ModuleVersionViewV1:
      additionalProperties: false
      properties:
        module_key:
          maxLength: 128
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*){1,7}$"
          title: "Module Key"
          type: "string"
        display_name:
          maxLength: 128
          minLength: 1
          title: "Display Name"
          type: "string"
        version:
          maxLength: 64
          minLength: 5
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          title: "Version"
          type: "string"
        state:
          enum:
            - "draft"
            - "validation_failed"
            - "validated"
            - "lab_accepted"
            - "published"
            - "deprecated"
            - "revoked"
          title: "State"
          type: "string"
        recipe:
          "$ref": "#/components/schemas/ModuleVersionViewV1__EndpointRecipeModuleSpecV1"
      required:
        - "module_key"
        - "display_name"
        - "version"
        - "state"
        - "recipe"
      title: "ModuleVersionViewV1"
      type: "object"
    ModuleVersionViewV1__EndpointRecipeModuleSpecV1:
      additionalProperties: false
      description: "A declarative, no-branching recipe expanded only by Endpoint Platform."
      properties:
        schema_version:
          const: "endpoint_recipe_module_v1"
          title: "Schema Version"
          type: "string"
        module_key:
          maxLength: 128
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*){1,7}$"
          title: "Module Key"
          type: "string"
        supported_platforms:
          items:
            enum:
              - "linux_amd64"
              - "windows_amd64"
            type: "string"
          maxItems: 2
          minItems: 1
          title: "Supported Platforms"
          type: "array"
        inputs:
          items:
            "$ref": "#/components/schemas/ModuleVersionViewV1__ModuleRecipeInputV1"
          maxItems: 8
          title: "Inputs"
          type: "array"
        steps:
          items:
            "$ref": "#/components/schemas/ModuleVersionViewV1__EndpointRecipeStepV1"
          maxItems: 8
          minItems: 1
          title: "Steps"
          type: "array"
      required:
        - "schema_version"
        - "module_key"
        - "supported_platforms"
        - "steps"
      title: "EndpointRecipeModuleSpecV1"
      type: "object"
...
    ModuleVersionViewV1__ModuleRecipeInputV1:
      additionalProperties: false
      properties:
        name:
          maxLength: 64
          minLength: 1
          pattern: "^[a-z][a-z0-9_]{0,63}$"
          title: "Name"
          type: "string"
        value_type:
          enum:
            - "string"
            - "integer"
          title: "Value Type"
          type: "string"
      required:
        - "name"
        - "value_type"
      title: "ModuleRecipeInputV1"
      type: "object"
...
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    ModuleOperationDetailEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/ModuleOperationDetailV1"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "ModuleOperationDetailEnvelope"
    ModuleOperationEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/ModuleOperationV1"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "ModuleOperationEnvelope"
...
    ModuleVersionViewEnvelope:
      properties:
        data:
          "$ref": "#/components/schemas/ModuleVersionViewV1"
      additionalProperties: false
      type: "object"
      required:
        - "data"
      title: "ModuleVersionViewEnvelope"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
  securitySchemes:
    AgentBearer:
      type: "http"
      scheme: "bearer"
    ServiceBearer:
      type: "http"
      scheme: "bearer"